requirements.txt
absl-py==2.3.1
ale-py==0.11.2
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
grpcio==1.73.1
gymnasium==1.1.1
ImageHash==4.3.2
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.9.79
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
pyboy==2.6.0
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
PySDL2==0.9.17
pysdl2-dll==2.32.0
python-dateutil==2.9.0.post0
pytorch-triton==3.4.0+gitae848267
pytz==2025.2
PyWavelets==1.8.0
rich==14.0.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
stable_baselines3==2.6.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.9.0.dev20250715+cu129
torchaudio==2.8.0.dev20250715+cu129
torchvision==0.24.0.dev20250715+cu129
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
